fix: change key to use recipe id
